clean inspections move Platform.py
